Resovling the crash due to geopositions
I've removed the Microsoft.Azure.DocumentDB.Core dependancy from the mobile projects as this feels unneeded for what we're doing (our mobile solution shouldn't care or know about our data storage technology).

To get around removing this whilst keeping the existing app working, I've created a new GeoPoint class which has a Coordinates property which we map to Long and Lat for use in the app.

We should ensure that we only use Long and Lat moving forward and to push that, I've marked the property as deprecated.